Move cli scripts to cmd folder
