+++ deafrica_conflux/__main__.py
@@ -1,3 +1,3 @@
-from .cli import main
+from .cmd.cli import main
 
 main()
