+++ deafrica_conflux/cmd/run_from_queue.py
@@ -4,9 +4,8 @@ import datacube
 import logging
 from rasterio.errors import RasterioIOError
 
-from ._cli_common import main, logging_setup
-from ._vector_file_utils import get_crs, guess_id_field, load_and_reproject_shapefile
-from .plugins.utils import run_plugin, validate_plugin
+from .common import main, logging_setup, get_crs, guess_id_field, load_and_reproject_shapefile
+from ..plugins.utils import run_plugin, validate_plugin
 
 import deafrica_conflux.db
 import deafrica_conflux.io
