+++ deafrica_conflux/cmd/run_one_scene.py
@@ -3,7 +3,7 @@ import datacube
 import logging
 from rasterio.errors import RasterioIOError
 
-from ._cli_common import main, logging_setup
+from .common import main, logging_setup
 from ._vector_file_utils import get_crs, guess_id_field, load_and_reproject_shapefile
 from .plugins.utils import run_plugin, validate_plugin
