+++ deafrica_conflux/cmd/stack.py
@@ -1,6 +1,6 @@
 import click
 
-from ._cli_common import main, logging_setup
+from .common import main, logging_setup
 
 import deafrica_conflux.stack
